+++ src/handlers/LocationHandler/index.js
@@ -1,12 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import { AppState } from 'react-native';
 
 export class LocationHandler extends React.Component {
   constructor(props) {
     super(props);
 
-    this.getLocationPermission = this.getLocationPermission.bind(this);
+    this.handleAppStateChange = this.handleAppStateChange.bind(this);
+    this.setAppState = this.setAppState.bind(this);
+    this.getLocation = this.getLocation.bind(this);
+
+    this.state = {
+      appState: null,
+    };
   }
 
   static get propTypes() {
@@ -16,10 +23,33 @@ export class LocationHandler extends React.Component {
   }
 
   componentDidMount() {
-    this.getLocationPermission();
+    this.getLocation();
+
+    AppState.addEventListener('change', this.handleAppStateChange);
   }
 
-  getLocationPermission() {
+  componentWillUnmount() {
+    AppState.removeEventListener('change', this.handleAppStateChange);
+  }
+
+  handleAppStateChange(nextAppState) {
+    const { appState } = this.state;
+
+    // If the app is foregrounded after being backgrounded
+    if (nextAppState === 'active' && appState && appState.match(/inactive|background/)) {
+      this.getLocation();
+    }
+
+    this.setAppState(nextAppState);
+  }
+
+  setAppState(appState) {
+    this.setState({
+      appState,
+    });
+  }
+
+  getLocation() {
     const { dispatch } = this.props;
 
     dispatch({